+++ client/src/components/DataFiles/DataFilesTable/DataFilesTable.js
@@ -142,7 +142,7 @@ const DataFilesTablePlaceholder = ({ section, data }) => {
   if (filesLength === 0) {
     return (
       <div className="h-100 listing-placeholder">
-        <SectionMessage type="warn">
+        <SectionMessage type="warning">
           No files or folders to show.
         </SectionMessage>
       </div>
@@ -190,6 +190,7 @@ const DataFilesTableRow = ({
         index={row.index}
         onClick={onClick}
         onKeyDown={onKeyDown}
+        data-testid="file-listing-item"
       >
         {row.cells.map(cell => {
           return (