+++ client/src/components/DataFiles/DataFilesToolbar/DataFilesToolbar.test.js
@@ -10,7 +10,7 @@ const mockStore = configureStore();
 expect.extend({ toHaveClass });
 describe('ToolbarButton', () => {
   const store = mockStore({});
-  it('render button with correct text', () => {
+  it('renders button with correct text', () => {
     const { getByText, getByRole, getByTestId } = renderComponent(
       <ToolbarButton text="Rename" iconName="rename" onClick={() => {}} />,
       store,
@@ -24,7 +24,7 @@ describe('ToolbarButton', () => {
 });
 
 describe('DataFilesToolbar', () => {
-  it('render necessary buttons', () => {
+  it('renders necessary buttons', () => {
     const { getByText, queryByText } = renderComponent(
       <DataFilesToolbar scheme="private" api="tapis" />,
       mockStore({
@@ -46,10 +46,77 @@ describe('DataFilesToolbar', () => {
     expect(getByText(/Trash/)).toBeDefined();
     expect(queryByText(/Make Public/)).toBeFalsy();
   });
-});
 
-describe('DataFilesToolbar', () => {
-  it('render Make Public button', () => {
+  it('does not render unnecessary buttons in Community Data', () => {
+    const {getByText, queryByText} = renderComponent(
+      <DataFilesToolbar scheme="community" api="tapis" />,
+      mockStore({
+        workbench: { config: {
+          extract: '',
+          compress: ''
+        } },
+        files: { selected: { FilesListing: [] } },
+        listing: { selected: { FilesListing: [] } },
+        systems: systemsFixture
+      }),
+      createMemoryHistory()
+    );
+
+    expect(queryByText(/Rename/)).toBeFalsy();
+    expect(queryByText(/Move/)).toBeFalsy();
+    expect(getByText(/Copy/)).toBeDefined();
+    expect(getByText(/Download/)).toBeDefined();
+    expect(queryByText(/Trash/)).toBeFalsy();
+    expect(queryByText(/Make Public/)).toBeFalsy();
+  });
+
+  it('does not render unnecessary buttons in Public Data', () => {
+    const {getByText, queryByText} = renderComponent(
+      <DataFilesToolbar scheme="public" api="tapis" />,
+      mockStore({
+        workbench: { config: {
+          extract: '',
+          compress: ''
+        } },
+        files: { selected: { FilesListing: [] } },
+        listing: { selected: { FilesListing: [] } },
+        systems: systemsFixture
+      }),
+      createMemoryHistory()
+    );
+
+    expect(queryByText(/Rename/)).toBeFalsy();
+    expect(queryByText(/Move/)).toBeFalsy();
+    expect(getByText(/Copy/)).toBeDefined();
+    expect(getByText(/Download/)).toBeDefined();
+    expect(queryByText(/Trash/)).toBeFalsy();
+    expect(queryByText(/Make Public/)).toBeFalsy();
+  });
+
+  it('does not render unnecessary buttons in Google Drive', () => {
+    const {getByText, queryByText} = renderComponent(
+      <DataFilesToolbar scheme="private" api="googledrive" />,
+      mockStore({
+        workbench: { config: {
+          extract: '',
+          compress: ''
+        } },
+        files: { selected: { FilesListing: [] } },
+        listing: { selected: { FilesListing: [] } },
+        systems: systemsFixture
+      }),
+      createMemoryHistory()
+    );
+
+    expect(queryByText(/Rename/)).toBeFalsy();
+    expect(queryByText(/Move/)).toBeFalsy();
+    expect(getByText(/Copy/)).toBeDefined();
+    expect(queryByText(/Download/)).toBeFalsy();
+    expect(queryByText(/Trash/)).toBeFalsy();
+    expect(queryByText(/Make Public/)).toBeFalsy();
+  });
+
+  it('renders Make Public button', () => {
     const { getByText } = renderComponent(
       <DataFilesToolbar scheme="private" api="tapis" />,
       mockStore({