+++ client/src/redux/reducers/workbench.reducers.js
@@ -1,8 +1,9 @@
 export const initialState = {
-  loading: false,
+  loading: true,
   config: {},
   portalName: '',
-  setupComplete: window.__INITIAL_SETUP_COMPLETE__
+  setupComplete: window.__INITIAL_SETUP_COMPLETE__,
+  hideDataFiles: false
 };
 
 export default function workbench(state = initialState, action) {