+++ CHANGELOG.md
@@ -7,6 +7,70 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.9.0] - 2021-09-01: Shared Workspace Updates; Cortal Icons
+
+### Changed
+
+- FP-1142: Distinguish types of message text constants (#445)
+- FP-1153: Shared Workspaces Icon (#466)
+- FP-223: Add Custom Close Icon to Modals (#465)
+- FP-635: Remove Font Awesome Icons (#469)
+- Reduce Body Min-* Verbosity (#421)
+- FP-1089: Update ProjectId on collision (#459)
+- FP-1143: Convert underscore to hyphen in system ids (#461)
+- FP-1150: 150 max character limit for shared workspace title (#460)
+- FP-1171: Toggle data files [A2CPS] (#471)
+- FP-1128: Accommodate non-portal created Projects (#458)
+
+### Fixed
+
+- FP-1151: Shared workspace privilege bypass (#462)
+
+## [2.8.0] - 2021-08-10: Data Files Filter; Custom Settings for A2CPS
+
+### Added
+
+- FP-508: Data Files Filter (#455)
+- FP-195: Enable Django impersonate feature (#456)
+
+### Changed
+
+- FP-1004: Update Onboarding Step Description Text (#454)
+- FP-1118: Support no datafiles systems (#449)
+- FP-1117: Read only shared workspace setting (#450)
+- FP-1129: Toggle Apps and History Sections (#447)
+- FP-1134: Toggle Manage Team and Edit Description (#451)
+- FP-952: Replace Django Messages with Client Side Messages (#443)
+
+### Fixed
+
+- FP-1110: Shared workspace description character limit (#444)
+- FP-1133: Ensure that Frontera-related normal queue validation occurs only on Frontera (#446)
+- FP-1080: Fix search loading (#453)
+- FP-1139: Update allocations cache during login (#448)
+
+## [2.7.0] - 2021-07-13: Apps, Shared Workspaces Improvements
+
+### Added
+
+- FP-1060: Folder download message (#425)
+
+### Changed
+
+- FP-1027: Remove unrelated objects from app definition (#387)
+- FP-939: Display shared workspace description whitespace. (#435)
+- FP-997, FP-1041, FP-1042: Improve apps, add small queue support, and update execution systems when needed (#426)
+- FP-573: Prevent submitting unlicensed app (#438)
+- FP-1009: Apply ReadMore UI pattern to long Shared Workspace descriptions (#442)
+- FP-1031: Clarify Onboarding MFA Steps (#441)
+- FP-1007: Remove inactive toolbar buttons for public/community/google drive (#436)
+
+### Fixed
+
+- FP-1095: Adjust colors of links in Manage Account to match the website's theme. (#430)
+- FP-1108: Fix HTML App Loading (#437)
+- FP-1030: Fix breadcrumbs for application page (#434)
+
 ## [2.6.1] - 2021-06-15: Hotfix: Fix settings import
 
 ### Fixed
@@ -120,7 +184,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add some intitial steps to README (#367)
 - FP-895: Re-style site search file listing as a table (#353)
 - FP-950: Quick: Remove <Message type="info"> icon (#362)
-- FP-385: section comp welcome msg  patterns (#347)
+- FP-385: section comp welcome msg patterns (#347)
 - FP-898: onboarding system health checks (#345)
 - FP-908: Hide empty application tabs (#343)
 - Bump poetry version (#356)
@@ -390,7 +454,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2020-02-28
 v1.0.0 Production release as of Feb 28, 2020.
 
-[unreleased]: https://github.com/TACC/Core-Portal/compare/v2.6.1...HEAD
+[unreleased]: https://github.com/TACC/Core-Portal/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/TACC/Core-Portal/releases/tag/v2.9.0
+[2.8.0]: https://github.com/TACC/Core-Portal/releases/tag/v2.8.0
+[2.7.0]: https://github.com/TACC/Core-Portal/releases/tag/v2.7.0
 [2.6.1]: https://github.com/TACC/Core-Portal/releases/tag/v2.6.1
 [2.6.0]: https://github.com/TACC/Core-Portal/releases/tag/v2.6.0
 [2.5.0]: https://github.com/TACC/Core-Portal/releases/tag/v2.5.0